fix: copying text from a list item
Previosly, if you copied a text from within a list item, and pasted it
into a paragraph, the editor would split the paragraph and insert a list.

With this change, if you copy a text that's fully within a single list
item, it will modify the copied fragment to include only the paragraph,
so that if you paste it into another paragraph, it will insert the text
into the paragraph as expected by the user.